Support compacting records for ease of debugging
This adds a helper utility to strip out empty values when loading
a record. A large portion of many records is just hash keys with
empty array values, which make it hard to read results and use up
space...this allows you to trim down the record when parsing in
development so things are easy to read and reason about.

Also separates the initializers for CocinaRecord so that it's
clear how to create one from JSON vs. from a hash.