Add an interface for parsing related resources
This adds a simple subclass and helper method to access the
related resources on a CocinaRecord. Because this structured data
is basically identical to the description section in a normal record,
we get access to all of the existing methods "for free" on every
related resource.

Dataworks uses this to get the identifiers (e.g. DOIs) associated with
related records, like datasets with many linked versions.