Better handling for unencoded but compliant edtf dates
This adds a few patterns to handle well-formed edtf dates that
did not declare edtf as their encoding, which resulted in them
using a less-accurate parser and being given only year precision.

Now, if you have an unencoded date like 2019-01-01, it is correctly
detected as full day precision.

This also applies to dates in a date range, which didn't inherit
an encoding from their top-level date until now.